Use RLE on step hash to make arrays and ranges equal
Instead of hashing the values themselves, hash the first value and
differences between subsequent elements using run-length encoding.
Elements for which the - operator is not defined are hashed directly.

This allows for O(1) hashing of (some) ranges consistent with AbstractArrays,
which means they can now compare equal. Types which have a regular range
step have to use the new TypeRangeStep trait to enable O(1) hashing rather
than the O(N) AbstractArray fallback.

Apply the new trait to date ranges which have a regular step. Add tests
for the new behaviors.