Remove order::Ordering keyword argument to sort* methods where possible
This argument is only used internally and was not supposed to be documented
(which was the case in Julia 0.6). Remove it everywhere it is not needed, that
is in all functions except sortperm(), for which that argument is used by
sortrows() and sortcols() internally. Remove it for docstrings anyway.

Also remove the initialized keyword argument to sort(), which had no effect
and was not documented in Julia 0.6 either.